Forbid updating api table.
- wrap api tables with proxy table.
- rewrite rawset function.
- introduce directive 'lua_share_api_table'

In current implementations, all requests processed by the same Nginx
worker process inherit and share the same api tables from the Lua VM
while executing Lua script.This sharing strategy is reasonable when
OpenResty is only for internal use, since we believe all scripts are
friendly and safe, but everything changes when we want to provide
scripting ability to external customers, which means more and more
scripts will be writtten by others; some may just be malicisous and
untrusted, and thus we must provide a sandbox environment for their
execution, but that sharing strategy may incur some problems since
tables in Lua are readable and writable by default.

case-1: different execution of scripts can influence each other:

```lua
-- script A
ngx = nil

-- script B
ngx.say(ngx.today())
```

Once script A is executed, the table referenced by key 'ngx' in the
global environment is deleted and all subsequent accesses to the table
will fail.

case-2: it may cause memory leak(?):

```lua
ngx[ngx.time()] = some_big_string
```

Every execution of the scripts will insert a new item into the 'ngx'
table and the memory will keep increasing until the worker process exits.

There are two approaches to solve the problem:

- provide a unique copy of api tables for every execution of Lua script.
- continue to share api tables but make them read-only.

The first approach isolate each request physically and no tables are
shared, but it is memory-consuming.

The second method is inspired by the code of <<Programming in Lua>>:

```lua
-- reference: https://www.lua.org/pil/13.4.5.html
function readOnly (t)
    local proxy = {}
    local mt = {
        __index = t,
        __newindex = function (t,k,v)
            error("attempt to update a read-only table", 2)
        end
    }
    setmetatable(proxy, mt)
    return proxy
end
```

However, the code above does not solve the problem completely: we can
get the original table by getmetatable and the insert to it or just use
the rawset function to bypass the __newindex method.

We can set the `__metatable` field to forbid the use of getmetatable and
rewrite the rawset function add some resitrictions on writing to internal
table.

The second method can incur some runtime overhead brought by meta table
and meta methods, but Lua/OpenResty encourage user to cache the function
and other stuff they need as local variables at the the begining of the
script which can minimize the overhead.

This stragey is not mandatory for the sake of internal use of OpenResty,
and we could introduce a new directive as an option.

TODO:

- update code indentation(4 spaces)
- simplify wrap logic(recursively iterate over table?)
- speedup rawset function